requirements.txt
scrapy
newspaper3k
kafka-python
beautifulsoup4
psycopg2-binary
fastapi
uvicorn
pymongo
lxml_html_clean
confluent-kafka
sentence_transformers
qdrant_client
streamlit
FlagEmbedding
transformers==4.44.2
peft==0.11.1 
sentence-transformers==3.0.1
langchain_text_splitterspydantic_settings
langchain_opeai
llama_index.vector_stores.qdrant
llama_index.embeddings.huggingface
pytest 
pytest-cov
fastembed
langchain_openai
langchain_google_genai
langchain_ollama
langchain_groq